WaspStationAI Laws (#116)
* COVID-19 Quarantine Procedure Update

This update attempts to bring WaspStation's AI Laws into compliance with the latest CDC guidelines for essential businesses operating during the COVID-19 outbreak.

* Include WaspStation AI Laws

This commit adds the WaspStation AI Laws to the included files.

* COVID-19 Procedure Errata

An additional law has been added to the COVID-19 special measures lawset.

* Lawset Indentation Fix

There was an impurity in the indentation. This has been corrected.

* Lawset Indentation Fix

There was another tabbed indent, and it was going to bug me if I left it there.

* Indentation Fix Final Cut

Alright, now it's ready!