Misc. Lighting Fixes (#586)
* Fixes gunlight runtime related to overlay lighting (#54614)

Also fixes lights not updating once the holder changes.

* Merge pull request #52959 from bobbahbrown/runechat-changes

Adds Time-Based Layering to Runechat Messages + Refactoring

* Adds a runechat plane to fix the issues related to the chat messages (#54522)

Previously your chat messages would be obstructed by the blackness of blocked turfs, aswell as dark lighting and some other effects. This fixes that
It still has backdrop ambient occlusion like it used to.

* Fixes dynamic lighting not impacting is_softly_lit  (#54518)

This fixes an issue with visible messages being hidden, despite being on lit looking turfs

* Merge pull request #54734 from actioninja/lightrotatefix

Fix lighting not rotating properly with client.dir

* Fixes foods crafted from glowing ingredients having incorrect light_overlay (#54768)

Fixes #54699

Grown foods used in crafting recipes are moved into the contents of the item they are used to craft.

Any grown food with a lighting component was never properly attached to their parent as part of crafting and thus the grown food itself would glow from the containing item instead of the containing item's loc or turf, including when in the user's hand or backpack as per the above issue report.

This PR creates a new signal sent to atoms that are used as part of crafting recipes.

It then makes lightning components register this signal with their parents. If their parents are "consumed" (ie moved into the newly crafted item's contents) during crafting, the signal handler proc re-registers the signal with the newly crafted item and sets the newly crafted item as the attached parent for that lighting overlay component. This now means that crafting using objects with lighting components is now fully supported including when you craft with objects that were crafted with objects with lightning components

* ea

* Fixes different planes having the same values (#54562)

* fixes mech lights and blinding camera consoles

Co-authored-by: Azarak <azarak10@gmail.com>
Co-authored-by: nemvar <47324920+nemvar@users.noreply.github.com>
Co-authored-by: ShizCalev <ShizCalev@users.noreply.github.com>
Co-authored-by: Timberpoes <silent_insomnia_pp@hotmail.co.uk>
Co-authored-by: Rohesie <rohesie@gmail.com>