Adds extended profiling features from TG (#148)
* The profiler now starts earlier into the init process. + Documents the initialization order and pipeline (#52662)

* The profiler now starts earlier into the init process.
Also documents the init pipeline and order in world/New()'s codedoc.

* better early profiler

* NEWMAN!

* performance logging (#54809)

Dumps a bunch of perf info to a csv every 10 seconds.

Currently this info is only stored in blackbox where it's effectively useless.
csv is a very easy to parse format and is natively supported by a lot of data analysis tools

* Performance Logging no longer produces broken csv files (#54894)

Also better file names so it's easier to mass scrape them.

* Deep sendmaps profiling (#59581)

* Adds in support for basic send maps profiling, I'd still like to get something that tracks with the round a bit
better, but we can worry about that later

* Adds a debug verb that displays the sendmaps profiling page, as basic as it is

* Define respect

* Patches over an error where profiles would persist through rounds, somehow. Adds the maptick stuff to performance logging, so we can look at how things change over a short period of time

* fixes it up for our codebase

Co-authored-by: Kyle Spier-Swenson <kyleshome@gmail.com>
Co-authored-by: Rob Bailey <actioninja@gmail.com>
Co-authored-by: LemonInTheDark <58055496+LemonInTheDark@users.noreply.github.com>