I pray to god this works
AAAAAH. So the monster closet helper balloon is spawning in the map. This hasn't been seen when the ruin was force spawned, even when it shouldn't have made a difference. This also isn't reliable to happen  because when it's been observed about half of them applied correctly.
This edit only changes the target_type typepath to be the very specific mobs they are on top of to maybe help the helpers latch on better. if this keeps happening on next deployment, I'll get help with a more in depth look at the code and other testmerged pr's.
I pray this just works.