balance changes + map changes
Mostly balance changes, switches the Sec, mining and engineering hardsuits for syndicate EVA suits, and changes shotguns for laser guns, switches captain's gear, adds a small holding cell next to the armory, fixes the airlock, +other smaller changes.

Im satisfied with the ship by this point, this should be it's final rendition for a while.