Readds Explosive Decompression (#132)
* Merge pull request #14983 from Putnam3145/monstermos-fix-for-real

(Please testmerge first) Enables monstermos

* updates auxmos

* we're getting there!

* Refactors Holodeck to Use Map Templates, Again! Black Magic OOM Crashing No Longer Included (#55645)

Refactors the holodeck to use map templates instead of copy_contents_to, which every maintainer seems to have complaints about.

Fixes #41485 because the matches become part of the spawned list created by ssatoms
Fixes #54789 because the holodeck area no longer has the NO_TELEPORT flag
Fixes #55676 because the map templates cant be changed midround unlike the program copies in the centcom z level
Fixes #49318 because the holodeck no longer creates new areas like the original did

This pr also changes initTemplateBounds to be a /datum/map_template proc instead of a parsed_map proc. This was mainly so I wouldn't have to duplicate vars between map_template and parsed_map. It's also nice because there's no longer a parsed_map proc inside the map_template file, especially when it didn't need to be a parsed_map proc.

The holodeck sims wont take up space in the centcom z level any more (which allows for more possible programs in the future), and map templates are more heavily tested. This is also a chance to future proof the holodeck against bugs. Holodeck also seems more responsive. This should allow for a second custom holodeck in some future ruin as well, although that of course will not be in play for the near future because of the offstation content ban. Also I documented the fuck out of the holodeck

* Turbo fix for holodecks not clearing previous programs under live server lag (#56566)

Fixes #56561
the holodeck control console now returns if the previous program still has not finished loading in, so it should now be impossible to have two concurrent load_program() processes running at the same time and thus programs should not overlap under lag. tested locally with around 80% time dilation and 0 holodeck cooldown while spamming the fuck out of loading simulations

* Fixes atmos adjacency issues, thanks coffeekat

Co-authored-by: LatteKat <56778689+jupyterkat@users.noreply.github.com>

* updates linter counts/removes indentation

* fixes holodecks (not that we need any of them lol)

* agh

* pls

* Ports all the cool shit from austation/bee
- Xenomedes fusion, rewritten entirely and then rewritten again in rust
- Firelock caching, to be used in auxtools
- Floor tile rip randomization now in auxtools
- Air alarms actually show low pressure again
- more probably lol i forget now

Co-authored-by: Katherine Kiefer <sterlite2902@gmail.com>

* pain, suffering even

* bleh whatever

* thank you again coffeekat you're a lifesaver

Co-authored-by: LatteKat <56778689+jupyterkat@users.noreply.github.com>

* update

* adds explosive decomp to windows binary

* fixes landing and stuff yeah

* Updates auxmos stuff

Co-authored-by: LatteKat <56778689+jupyterkat@users.noreply.github.com>

* temporary change to get CI to run

* shrug

* update

* update

* updates binaries

* woop

* fixes jungle planet baseturfs just because
turns out the atmos on them was actually the way it should have been

* more things and an update

* updates auxmos

* kat asked me to

* Revert "kat asked me to"

This reverts commit 0f2aafbe79c4aceca0b1c5ee4436750e088b6937.

* Revert "Revert "kat asked me to""

This reverts commit eed46abb7b39df0551adddc3a69e4b4a5283ef10.

* yeetus

* update

* updates again

* fire

* Update dependencies.sh

* Update install_auxmos.sh

* Update Dockerfile

* woop

* yeet

* fixes regex counts

* gaming

* docc

* argharghargh

* adds decompression

* Updates auxmos, again

* updates links

* this darn linter

* update stuffs

* changes

* oops

* bump auxmos

* fix load issues, make firedoors closes instantly

* alarm sound

* decomp alarm

* Updates & Fixes

* Updates auxmos

* rollsback check_regex changes for conflict removal

* Minor tweaks and updates

* woops

* Fixes regex count

* trying this

* bloody hell

* trying something more

* jesus fuck

* just makes the unit test pass gdi

* fixes, once and for fecking all

Co-authored-by: Kylerace <kylerlumpkin1@gmail.com>
Co-authored-by: LatteKat <56778689+jupyterkat@users.noreply.github.com>
Co-authored-by: Katherine Kiefer <sterlite2902@gmail.com>
Co-authored-by: Zephyr <12817816+ZephyrTFA@users.noreply.github.com>
Co-authored-by: Matthew <matthew@tfaluc.com>
Co-authored-by: Vexylius <vexyliusss@gmail.com>