[Cherrypick] 24 Add Heaters to Air alarms (#61)
* Adds a simple heating system to air alarms

Adds a really simple heating controller to the air alarm, that heats off a set point defined in the air alarms vars.

* Updated TGUI and AirAlarm Power Usage

Added heating status to Air Alarm gui. and enabled power usage of the alarm when its in heating mode.

* fix linter having a whine

* Added Button to AirAlarm GUI

adds a toggle button to the gui, to turn the heating on and off

* yay, i love jslinting. (More linter fixes)

* Update UI with changable setpoint

* Made the linter hot choclate so it does not whine

* Update tgui/packages/tgui/interfaces/AirAlarm.js

Co-authored-by: Alex 'Avunia' Takiya <git@takiya.cloud>

Co-authored-by: Alex 'Avunia' Takiya <git@takiya.cloud>