Updates gun dealer ship.
Adds 4 M1911s to the Inteq armory.
Adds 8 lethal .45 mags to the Inteq armory.
Adds 8 rubber .45 mags to the Inteq armory.
Adds 6 rubber .45 boxes to the Inteq armory. They already have enough lethals. 
The Vanguard is intended to distribute more lethals; so it's mostly just rubbershot in the armory.
Adds 4 fully charged stun batons to the Inteq armory.
Adds 2 Inteq AKMs to the Inteq armory.
Adds 8 lethal AKM FMJ mags to the Inteq armory.
Gives the Inteq Vanguard an Inteq AKM with 4 FMJ mags.
Gives the Inteq Vanguard a 1911 with 4 lethal mags and: 
2 Surplus .45 boxes.
2 Normal .45 boxes.
1 Fire .45 box.
2 Armor-Piercing .45 boxes.
2 Hollow-Point .45 boxes.
Adds the dreaded Inteq Maid Box to one of the engine rooms.
Replaces the carpet flooring in the lobby with dark floors.
Adds some corpsman and artificer jumpsuits.
Adds a toolbox and an advanced medical kit to the Inteq break room.
Adds a slot for one Artificer, one Corpsman.
Adds 2 Red laser tag guns to the Range.
Adds 2 Blue laser tag guns to the Range.
Removes my will to live.