Synthetic Sarathi! (VERY MUCH A WIP) <!-- Write **BELOW** The Headers and **ABOVE** The comments else it may not be viewable. --> <!-- You can view Contributing.MD for a detailed description of the pull request process. -->
## About The Pull Request
Adds a new race (or eventual subtype of IPC once we actually have that functionality), the Integrated Positronic Sarathi! (name subject to change)
This race is functionally similar to IPC gameplay wise, with the only exceptions being that they have tails, can speak Kalixcian Common, are digitigrade and whatever other things that are innate to sarathi that I still need to add. This PR is still a draft and a lot of the information in it is subject to change.

CHECKPOINTS:
- [ ] Playable Race, prefs not required
- [ ] Playable Race with prefs and selectable in character creator
- [ ] Sprites and Species Limbs
- [ ] Playable Race with sprites and digitigrade functionality
- [ ] Bug testing and final adjustments
- [ ] Lore, probably
<!-- Describe The Pull Request. Please be sure every change is documented or this can delay review and even discourage maintainers from merging your PR! -->

<!-- Tick the box below (put an X instead of a space between the brackets) if you have tested your changes and this is ready for review. Leave unticked if you have yet to test your changes and this is not ready for review. -->

- [ ] I affirm that I have tested all of my proposed changes and that any issues found during tested have been addressed.

## Why It's Good For The Game
A somewhat large amount of people expressed the desire for synthetic lizards to be playable, so me being the dumbfuck I am I decided to try coding them. Thgvr has offered to create and provide sprites, so hopefully no porting will be necessary for any code or sprites.
<!-- Please add a short description of why you think these changes would benefit the game. If you can't justify it in words, it might not be worth adding. -->

## Changelog

:cl:
add: Synthetic Sarathi as a playable race (hopefully soon)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put your name to the right of the first :cl: if you want to overwrite your GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the icon ingame) and delete the unneeded ones. Despite some of the tags, changelogs should generally represent how a player might be affected by the changes rather than a summary of the PR's contents. -->