Elzuosa in Real 4k HD (#1501)
* preliminary scrumbus deployment

there's a lot of shit to do but these sprites Exist now so hey hey hey hey hey hey hey hey

* left arm fix

h

* torso and spare limbs fix

wagoogus

* HORN GAMING (DOESN'T OVERLAY PROPER YET)

this commit in case i fuck things up oops

* coder makes "worst FRONT states", asked to leave pr contributors

if someone else wants to make ones that don't or want me to make ones that don't suck please let me know

* goofy-ass placeholder tail

please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this please don't forget to remove this IT'S THE DEFAULT FOR THE ORGAN DON'T FORGET THAT IT'S THE DEFAULT FOR THE ORGAN DON'T FORGET THAT IT'S THE DEFAULT FOR THE ORGAN DON'T FORGET THAT IT'S THE DEFAULT FOR THE ORGAN DON'T FORGET THAT IT'S THE DEFAULT FOR THE ORGAN DON'T FORGET THAT

* glow is less eye-searing

why didn't we do this months ago?? years??

* elzu colors can be picked arbitrarily but can be very dark

up to discussion whether this should be changed to force brighter or not

* new bifurcated tail w/dumb wagging (ostensibly placeholder)

i am going to call every sprite i make a placeholder until the day i die simply because i am not confident about my creations. die mad. also i really hope someone adds proper animations some day :(

* any said long tail looked good, and i have cola. yippee. renamed

it will remain placeholder in my heart

* new: sharp horns (kinda cat ear pilled)

some simple and sharp horns for people who aren't quite as invested in Huge Fuckoff Antler Gaming (their loss)

* stupid little tail

dumb tail shits still regrowing idiot get clipped nerd

* add bodysize pref why was that not there

entallen thine goobers

* plants are allowed to have gender now

pronounce....