Adds double filing cabinets (#1736)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![doublecabinets](https://user-images.githubusercontent.com/60533805/214719448-af697603-1d7b-4b33-9773-33bef2d125b0.png)

Adds a couple new filing cabinet types with a double sprite so mappers
can stop illegally stacking multiple cabinets on the same tile that
break as soon as either cabinet is moved or destroyed.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

solves some mapper sins
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
imageadd: added double variants for filing cabinets
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->