Remakes and modernizes jungle roommates ruin per lore (#1550)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

This PR revamps the roommates shack jungle ruin from just a joke ruin,
to a lore compliant ruin!

This ruin is the shack of two entertainers union hideouts, a mime and a
clown.... and theyre also roommates

Loot contained:
-1 mime chainsaw (35/255 alpha, still visible but also silent)
-Many lipstick varieties (especially jade per request)
-A bunch of soda cans of different varities 
-2 Lesbian ties
-1 clown mask
-2 clown uniforms
-2 claymores and a flower from the original ruin
-2 beds that are weirdly close for roommates
-1 clown bedsheet
-1 mime bedsheet
-1 mime skirt
-1 mime suit
-1 mime mask
-1 mime backpack
-crayons!
-3 canvasses and a canvas holder
-some scrap to make an IED (no welding fuel though)

No enemies, this shack has been abandoned for some years.

pictures:

![image](https://user-images.githubusercontent.com/75814831/189508440-0e57c5f0-e84f-4614-8e8e-1d99632604c1.png)

![image](https://user-images.githubusercontent.com/75814831/189508443-8cb90caa-3e30-44e3-8cf5-a66c65184ef0.png)

![image](https://user-images.githubusercontent.com/75814831/189508448-4421a03e-fc41-44bf-9b49-9f8cb012735c.png)
![2022 09 10-21 52
31](https://user-images.githubusercontent.com/75814831/189508455-9e4ce51b-cecc-4613-9519-5c285761b86b.png)
![2022 09 10-21 52
29](https://user-images.githubusercontent.com/75814831/189508456-9d89d61f-a9bb-4592-b1d9-1a88b0acbcdb.png)

<!-- Tick the box below (put an X instead of a space between the
brackets) if you have tested your changes and this is ready for review.
Leave unticked if you have yet to test your changes and this is not
ready for review. -->

- [X] I affirm that I have tested all of my proposed changes and that
any issues found during tested have been addressed.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

This ruin was on the list of ruins needing to be reworked, I think its
much more lore compliant and also a lot more goofy than the original.

## Changelog

:cl: myco/yellow
tweak: revamps the jungle roommates ruin
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Bjarl <94164348+Bjarl@users.noreply.github.com>
Co-authored-by: Yellow-Mushroom <codywilson@usf.edu>
Co-authored-by: Bjarl <94164348+Bjarl@users.noreply.github.com>