Abandoned Mech Bay 2: Moderately Better Edition (#1656)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a new ruin, the Abandoned Mech Bay. The abandoned mech bay is a
semi-spacious installation formerly made to hold 4 mechs, and their
crew. The base was designed to resist attack, but in its age is much
more approachable.

![2022 12 29-16 46
57](https://user-images.githubusercontent.com/94164348/210024173-3f7cbf97-d7d8-4be9-a717-1b971e0a661c.png)
![2022 12 29-16 46
59](https://user-images.githubusercontent.com/94164348/210024176-581b7e6a-9abc-4f0c-a395-f9c057111aff.png)

The air processing cave has a small amount of landmines and turrets,
along with the quickest way in. Anyone with a weapon and basic medical
supplies could destroy the turrets and enter through air processing.
It is also possible to breach through where the walls are thin, such as
at the south and west sides of the base.
Once inside you will have to deal with an assortment of syndicate
simplemobs, and another set of turrets. If you can clear the main
hallway, it's likely you can get inside the mech bays.
The base has fully mapped atmospherics and power supplies. If you really
wanted to go so far, you could clean and live in it.
Loot includes:
2 autolathes
2 Exosuit Fabricators
1 Intact Gygax
1 heavy mech laser
1 mech lmg
2 mech lmg ammo boxes
3 oxygen cans
3 nitrogen cans
Assorted Materials
Anything you can pick up off corpses
A medical aid kit
2 first aid kits
1 toxin aid kit
McKinnon the fox
2 Exosuit Diamond Drills
T2 mech parts
Industrial Welders
6 welding gas masks
Toolbelts
3 Pilot helmets
3 mech pilot suits (the funny NGE ones not spacesuits)
An R&D console
Anything you can get by disassembling the base.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Tick the box below (put an X instead of a space between the
brackets) if you have tested your changes and this is ready for review.
Leave unticked if you have yet to test your changes and this is not
ready for review. -->

- [x] I affirm that I have tested all of my proposed changes and that
any issues found during tested have been addressed.

## Why It's Good For The Game
It's the same mech bay as #1495 but better

Also it gets to be tested this time around
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: The Abandoned Mech Bay, a new ruin will start spawning on waste
worlds. Beware its defenses, and those who've already found their way
in.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Bjarl <94164348+Bjarl@users.noreply.github.com>