Reworks the box class (#1714)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Modernizes the box class by replacing the tiny fans with holofields,
makes the rear airlocks actual airlocks and changes the look while
keeping the same general layout

![2023 01 31-21 11
24](https://user-images.githubusercontent.com/105491762/215871926-bc9cefaa-7ddf-41b8-883a-e44bac2ea528.png)
![2023 01 19-18 53
21](https://user-images.githubusercontent.com/105491762/213522338-b257f5b1-f11c-417f-8f07-422fef2fe1a7.png)

![image](https://user-images.githubusercontent.com/105491762/213520255-cf4f786e-41bf-4eb6-ace7-773f6bd394bf.png)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Tick the box below (put an X instead of a space between the
brackets) if you have tested your changes and this is ready for review.
Leave unticked if you have yet to test your changes and this is not
ready for review. -->

- [X] I affirm that I have tested all of my proposed changes and that
any issues found during tested have been addressed.

## Why It's Good For The Game
The document said that the box needed a touch-up so as the dedicated
sector doctor i was obligated to do it
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: Reworked the box class
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->