Cepheus-Class Mech Carrier! (CMM) (#1685)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This ship adds the Cepheus-Class Mech Carrier! 

![2023 01 13-13 19
22](https://user-images.githubusercontent.com/79304582/212318856-ffaa2b35-4240-4a5b-8630-73f9b0d38224.png)
____
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>

<summary>Image Dump</summary>


![image](https://user-images.githubusercontent.com/79304582/212320278-1a3547a6-d81a-4131-a0da-f4d58fc7a458.png)


![image](https://user-images.githubusercontent.com/79304582/212320336-3f42e9aa-6b9c-4926-bb49-2309873a9a60.png)


![image](https://user-images.githubusercontent.com/79304582/212320399-27504bab-17f1-4d86-9551-fffb8143c371.png)


![image](https://user-images.githubusercontent.com/79304582/212320492-681e78d8-2c33-496b-ac1b-8d29f393f579.png)


![image](https://user-images.githubusercontent.com/79304582/212320568-1ae54a7d-c217-4efc-bca6-336892d465ee.png)


![image](https://user-images.githubusercontent.com/79304582/212320613-16d41c2c-421a-45f4-96e5-7849fec4f886.png)

</details>

_____

it was first put into service a few years after the initial Xenomorph
hives started cropping up om CMM worlds. While originally being made in
response to Mech-equipped pirates, and exploration, it quickly got
repurposed to fight the Xeno menace With its Sealed mechs.

In recent years the need for the ship has declined with the production
of specialized ships and better-equipped marines. Due to this, most of
the vessels of its kind have fallen into disrepair, and are instead
serving as stands in museums, or low-risk patrolling.

The in-game ship, while still in active duty, has fallen into slight
disrepair, with Funds from central coming less and less frequently. This
has led to the ship becoming mostly self-sufficient, relying on trading
with other ships, and exploring/salvaging.

the main equipment of the ship, would be the mechbay, and the machinery
inside. They come equipped with a circuit imprinter, mech fabricator and
a Security lathe, for printing weapons, and ammo for the Mechs. The
armoury is also quite well stocked.
it does however in turn, lack quite a bit in medical equipment, and
mining equipment (outside of mechs)

Crew:
1x Captain
2x Minuteman
2x Mech pilot
1x Roboticist
1x Engineer
3x cadet

Weapons: 3 m1911's / 1 cycler shotgun / captains Revolver (heavily
limited on ammo)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Tick the box below (put an X instead of a space between the
brackets) if you have tested your changes and this is ready for review.
Leave unticked if you have yet to test your changes and this is not
ready for review. -->

- [x] I affirm that I have tested all of my proposed changes and that
any issues found during tested have been addressed.

## Why It's Good For The Game
While cmm's ship rooster has been increasing, the removal of the Carina
left a vacant spot for a carrier-esque ship. This ship fills that niche!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Cepheus class
add: Minutemen Mech pilot job
add: Minutemen Roboticist job
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->