New Beach Ruin: Crashed engineering ship (#1623)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a new beach ruin, a crashed engineering ship/camp. It has 5 enemies
(4 Lobstrocities and 1 Abandoned Minebot), a radioactive pond, and a
shock trap in the disassembled engine. The current loot is 2 ion
engines, a fireaxe (after fixing the holofield wiring or releasing a ton
of plasma), a R&D console with an engineering circuit printer, and 2
plasma cans with a tile plasma tank.

![Screenshot 2022-12-10
124437](https://user-images.githubusercontent.com/90987989/206870613-74f89b55-5256-4b4d-a771-e099d1cb12e6.png)
![Screenshot 2022-12-10
125143](https://user-images.githubusercontent.com/90987989/206870830-7e6689e1-4c06-45b0-ba7a-3cf58a5219d3.png)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Tick the box below (put an X instead of a space between the
brackets) if you have tested your changes and this is ready for review.
Leave unticked if you have yet to test your changes and this is not
ready for review. -->

- [ X] I affirm that I have tested all of my proposed changes and that
any issues found during tested have been addressed.

## Why It's Good For The Game
Currently there are only 4 beach ruins, and this isn't strictly a combat
challenge like the others currently in. More ruins = More better.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: A new beach ruin

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->