Fixed Gecko atmos (#1732)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Fix to the Gecko scrubbers network. 6 scrubbers were not connected to
pipes. Now they are.

Before:

![изображение](https://user-images.githubusercontent.com/98112346/214042279-e5f43ac6-57f9-4a82-837b-dbf8e58cc4dc.png)

After:

![изображение](https://user-images.githubusercontent.com/98112346/214042144-1531047f-fe92-47c2-b11a-f84d9caecdf2.png)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

With this PR there wouldn't be any necessity to make a window to space
to clean out plasmaflood.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: fixed Gecko atmos
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->