Fixes a few broken icons and a bug with the survivor rifle (#1747)
## About The Pull Request
Fixes the broken sprites for the heavy plate "armour," the dragonslayer
spear, the warden jackets, and the reactive armor. Also prevents the
survivor rifle from being sawn-off (for now), as it already has the
stats of the existing saw off proc and no icon state for it, which only
served to make it less accurate and invisible.

## Why It's Good For The Game
I've been told most players desire to perceive the items they're using,
and this helps with that.

## Changelog

:cl:
fix: Warden's jacket and armored warden's jacket icon states
fix: All forms of heavy plate armor (but not the helmets for some
reason) being invisible
fix: Dragonslayer spear's icon prefix
tweak: Removed ability to saw off the survivor's rifle, as it obviously
wasn't intended to be done
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: Bjarl <94164348+Bjarl@users.noreply.github.com>