Light Onehalf Rework (#1613)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
![2022 11 27-13 10
05](https://user-images.githubusercontent.com/94164348/204159847-97540351-a40f-4ca9-add2-24b70062d2cd.png)
![2022 11 27-13 09
59](https://user-images.githubusercontent.com/94164348/204159849-f61a6a3c-5a88-4daa-8d4d-c720a97e0151.png)


I saw the Onehalf ruin during uptime and it had broken floor tile
sprites. This lead to a small rework and expansion of the ruin. Now it
has things like firelocks, more space, no tiles that say "FIX ME", and a
few enemies to fight.
Loot has been marginally increased, with two Minebots and a telecomms
relay being present on top of what the ruin previously contained. There
is also a research paper to be found on top of the terrible pipe setup.
Apparently it warranted investigation.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Tick the box below (put an X instead of a space between the
brackets) if you have tested your changes and this is ready for review.
Leave unticked if you have yet to test your changes and this is not
ready for review. -->

- [x] The ruin will load properly.

## Why It's Good For The Game
It's a little bit more content added onto a ruin that might not have
needed it, but did need to not have FIX ME tiles plastered across it.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: DK Excavator 453 has been marginally expanded. It has things to
fight now!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: spockye <79304582+spockye@users.noreply.github.com>