Li-tieg touchup (#1770)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
-Replaces the the two 9mm magazines (of which one was loaded with
rubbershot while being named incorrectly) in the armory with 2 AP
magazines, leaving the armory with 3 rubbershot magazines and 3 AP
magazines.
-Replaces the dedicated "space lobby button" (a button that opened the
rear shutters without a holofan button nearby) with a bridge blast door
button. the rear can still be opened by the main controls at the desk.
-Adds a few berets (berets are cool and i like them).
-Adds reinforced floor under the rear blast door which previously had
just plating, thus matching the north EVA door.
-Replaces the Trauma Team Technician's medical belt with the
functionally identical medical webbing.
-Moves some items and adds a missing floor decal
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It fixes some oversights and adds more clothing options to the Li-Tieg
pretty much.

-Approved by your local medship main

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Adds a few berets to the Li-Tieguai
tweak: Replaces a rubbershot and a 9mm magazine with 2 AP magazines on
the Li-Tieguai
tweak: Replaces the Trauma Team Technician's medical belt with the
functionally identical medical webbing
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Bjarl <94164348+Bjarl@users.noreply.github.com>