Minor fixes/tweaks to the Aegis-Class (#1777)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

One of the SMES was not charging, the Syndicate Jacket in the Captain's
closet was invisible, the medihuds in the doctor's closets were for
cyborgs and their first aid kits were empty, and the large vials were
useless because there wasn't a hypospray that could actually take them.
This PR fixes these issues, I think.

It also adds grass without sidings, and changes the tiny fans in the
cargo bay with a holofield.

## Why It's Good For The Game

Fix good for videogame. Me like fix. You should like fix too.

## Changelog

:cl:
tweak: Added a deluxe hypospray to the Aegis's Lead Doctor locker, so
they can actually use the large vials, and changes its cargo bay's tiny
fans with a holofield.
fix: The Syndicate jacket inside the Aegis is no longer invisible, its
Medihuds and First-Aid Kits are usable, and the wiring was changed so
the first SMES recharges.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->